Move ToDoConfiguration to package
