+++ src/main/java/ru/team/todo/config/ToDoConfiguration.java
@@ -1,4 +1,4 @@
-package ru.team.todo;
+package ru.team.todo.config;
 
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
